graphql-api/requirements.txt
strawberry-graphql[fastapi]
uvicorn[standard]
python-jose[cryptography]
requests